fix(yukigo): corrijo manejo de entorno en `applyArguments` y `visitApplication`
